face-verification-service/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
opencv-python-headless
face-recognition
numpy
scipy
python-multipart
